backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "ChatPT backend -- clinical chat + RAG-powered Dedalus tools"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "convex>=0.7.0",
    "dedalus-labs>=0.2.0",
    "fastapi>=0.128.3",
    "langchain>=1.2.9",
    "langchain-core>=1.2.9",
    "pydantic>=2.12.5",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.34.0",
    # RAG / Dedalus tools
    "docling>=2.0.0",
    "fastembed>=0.4.0",
    "qdrant-client>=1.12.0",
    "groq>=0.11.0",
    "instructor>=1.5.0",
    "semantic-text-splitter>=0.18.0",
    "sentence-transformers>=3.0.0",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.15.0",
]
